refactored and modified test for send message endpoint
